Reworked the external API of TlvTable to allow removing entries. Refactor of Request/Response packets making all optional parameters nullable, thus allowing removing them. Added more utility code to help splitting long messages.

+++ AberrantSMPP/Packet/Response/SmppBindResp.cs
@@ -55,8 +55,7 @@ namespace AberrantSMPP.Packet.Response
 		{
 			get
 			{
-				return GetOptionalParamString((ushort)
-					Pdu.OptionalParamCodes.SC_interface_version);
+				return GetOptionalParamString(Pdu.OptionalParamCodes.SC_interface_version);
 			}
 			
 			set
@@ -66,8 +65,7 @@ namespace AberrantSMPP.Packet.Response
 				{
 					version = string.Empty;
 				}
-				SetOptionalParamString(
-					(ushort)Pdu.OptionalParamCodes.SC_interface_version, version);
+				SetOptionalParamString(Pdu.OptionalParamCodes.SC_interface_version, version);
 			}
 		}
